# GFBackend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Gator Forum Backend Server",
        "title": "Gator Forum Backend API",
        "termsOfService": "https://github.com/fongziyjun16/SE/tree/backend",
        "contact": {},
        "version": "1.0"
    },
    "host": "http://167.71.166.120:10010",
    "basePath": "/gf/api",
    "paths": {
        "/article/create": {
            "post": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need new article info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article Manage"
                ],
                "summary": "Create a new article",
                "parameters": [
                    {
                        "description": "Create New Article",
                        "name": "ArticleInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ArticleInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Create Successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters / Info Error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/article/delete/:id": {
            "get": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need new article id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article Manage"
                ],
                "summary": "Delete Article By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Delete Successfully no matter what",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/article/getone/:id": {
            "get": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article Manage"
                ],
                "summary": "Get One Article By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Article ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Get Successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.ArticleDetail"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters / Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/article/update": {
            "post": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need ID, Title, Content in article info only",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article Manage"
                ],
                "summary": "Update Article Title or Content By ID",
                "parameters": [
                    {
                        "description": "Update Article Info",
                        "name": "ArticleInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ArticleInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Update Successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters / Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/articlelike/create/:articleID": {
            "post": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need article id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article Like Manage"
                ],
                "summary": "User like Article",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "233333",
                        "name": "ArticleID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Create Successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters or Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/articlelike/delete/:articleID": {
            "get": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need article id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article Like Manage"
                ],
                "summary": "User like Article",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "233333",
                        "name": "ArticleID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Delete Like Successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.ArticleDetail": {
            "type": "object",
            "properties": {
                "CommunityName": {
                    "type": "string",
                    "example": "big bang theory"
                },
                "Content": {
                    "type": "string",
                    "example": "I love UF"
                },
                "ID": {
                    "type": "integer"
                },
                "NumComment": {
                    "type": "integer"
                },
                "NumFavorite": {
                    "type": "integer"
                },
                "NumLike": {
                    "type": "integer"
                },
                "Title": {
                    "type": "string",
                    "example": "Gator Forum"
                },
                "TypeName": {
                    "type": "string",
                    "example": "music"
                }
            }
        },
        "entity.ArticleInfo": {
            "type": "object",
            "properties": {
                "CommunityID": {
                    "type": "integer"
                },
                "Content": {
                    "type": "string",
                    "example": "I love UF"
                },
                "ID": {
                    "type": "integer"
                },
                "Title": {
                    "type": "string",
                    "example": "Gator Forum"
                },
                "TypeID": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiAuthToken": {
            "type": "apiKey",
            "name": "token",
            "in": "cookies"
        }
    }
}